Function to take percentage and move rocket
Implemented a function which takes a percentage (i.e. how far along the line a user has moved the slider) and converts this to a position relative to the length of the trajectory. This then moves and rotates the rocket to that position.

Integrated this function to work with the playback panel and to pause the animation if the user moves the slider during play.