Created Filtering Script
A script has been created for filtering, allowing the user to select a data value from the unity editor and set a min/max bounds. Currently the filtering is functional with the max bounds.

Also created a new function within the IATK class "CSVDataSource" to allow external class to access the DimensionData for each csv file.