Temporary fix for trajectory scaling bug
Altered NormaliseCol method in CSVDataSource class of IATK to set the minimum and maximum values of each x,y,z axis before normalisation. This ensures that each trajectory gets scaled to the same bounds.

Note: This is a temporary fix so the trajectories can be visualised properly. Further work will need to be done to make this dynamic.