Fix Trajectory Local Rotation + Runtime Errors
The dynamic creation of trajectories now correctly rotates and positions the trajectories according to the rotation and position of the parent visualisation object.

Additionally, upon program start the visualisation checks if there is a dataSource object connected via the Unity Editor before attempting to create a data object. This resolves two runtime errors.